Registering Int8/16/32 data types for Pad operator. Also integrating float kernel and registering it.
