+++ tensorflow/lite/micro/kernels/xtensa/pad_int8_int16_int32_float32.cc
@@ -1,0 +1,85 @@
+/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+#include "tensorflow/lite/kernels/internal/reference/pad.h"
+
+#include <string.h>
+
+#include "tensorflow/lite/c/builtin_op_data.h"
+#include "tensorflow/lite/c/common.h"
+#include "tensorflow/lite/kernels/internal/types.h"
+#include "tensorflow/lite/kernels/kernel_util.h"
+#include "tensorflow/lite/kernels/op_macros.h"
+#include "tensorflow/lite/micro/kernels/kernel_util.h"
+#include "tensorflow/lite/micro/kernels/xtensa/xtensa.h"
+#include "tensorflow/lite/micro/kernels/xtensa/xtensa_pad.h"
+#include "tensorflow/lite/micro/micro_log.h"
+
+namespace tflite {
+
+namespace {
+#if defined(HIFI3) || defined(HIFI4) || defined(HIFI5)
+TfLiteStatus EvalInt8(TfLiteContext* context, TfLiteNode* node) {
+  return PadEvalHifiInt8(context, node);
+}
+
+TfLiteStatus EvalInt16(TfLiteContext* context, TfLiteNode* node) {
+  return PadEvalHifiInt16(context, node);
+}
+
+TfLiteStatus EvalInt32(TfLiteContext* context, TfLiteNode* node) {
+  return PadEvalHifiInt32(context, node);
+}
+
+#if defined(INCLUDE_FLOAT_OPT)
+TfLiteStatus EvalFloat32(TfLiteContext* context, TfLiteNode* node) {
+  return PadEvalHifiFloat32(context, node);
+}
+#endif
+#endif
+}  // namespace
+
+TFLMRegistration Register_PAD_INT8() {
+#if defined(HIFI3) || defined(HIFI4) || defined(HIFI5)
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, EvalInt8);
+#else
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, PadReferenceEvalInt8);
+#endif  
+}
+
+TFLMRegistration Register_PAD_INT16() {
+#if defined(HIFI3) || defined(HIFI4) || defined(HIFI5)  
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, EvalInt16);
+#else
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, PadReferenceEvalInt16);
+#endif    
+}
+
+TFLMRegistration Register_PAD_INT32() {
+#if defined(HIFI3) || defined(HIFI4) || defined(HIFI5)  
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, EvalInt32);
+#else
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, PadReferenceEvalInt32);
+#endif    
+}
+
+TFLMRegistration Register_PAD_FLOAT32() {
+#if defined(INCLUDE_FLOAT_OPT) && (defined(HIFI3) || defined(HIFI4) || defined(HIFI5))
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, EvalFloat32);
+#else
+  return tflite::micro::RegisterOp(XtensaPadInit, XtensaPadPrepare, PadReferenceEvalInt32);
+#endif    
+}
+
+}  // namespace tflite
